# backend/api/game/v1/game_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.1
// source: api/game/v1/game.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationGameCancelMatch = "/api.game.v1.Game/CancelMatch"
const OperationGameGetMatchStatus = "/api.game.v1.Game/GetMatchStatus"
const OperationGameStartMatch = "/api.game.v1.Game/StartMatch"

type GameHTTPServer interface {
	CancelMatch(context.Context, *CancelMatchRequest) (*CancelMatchReply, error)
	GetMatchStatus(context.Context, *GetMatchStatusRequest) (*GetMatchStatusReply, error)
	StartMatch(context.Context, *StartMatchRequest) (*StartMatchReply, error)
}

func RegisterGameHTTPServer(s *http.Server, srv GameHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/game/start", _Game_StartMatch0_HTTP_Handler(srv))
	r.POST("/v1/game/cancel", _Game_CancelMatch0_HTTP_Handler(srv))
	r.GET("/v1/game/status/{matchid}", _Game_GetMatchStatus0_HTTP_Handler(srv))
}

func _Game_StartMatch0_HTTP_Handler(srv GameHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in StartMatchRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGameStartMatch)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.StartMatch(ctx, req.(*StartMatchRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*StartMatchReply)
		return ctx.Result(200, reply)
	}
}

func _Game_CancelMatch0_HTTP_Handler(srv GameHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CancelMatchRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGameCancelMatch)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CancelMatch(ctx, req.(*CancelMatchRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CancelMatchReply)
		return ctx.Result(200, reply)
	}
}

func _Game_GetMatchStatus0_HTTP_Handler(srv GameHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMatchStatusRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGameGetMatchStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMatchStatus(ctx, req.(*GetMatchStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetMatchStatusReply)
		return ctx.Result(200, reply)
	}
}

type GameHTTPClient interface {
	CancelMatch(ctx context.Context, req *CancelMatchRequest, opts ...http.CallOption) (rsp *CancelMatchReply, err error)
	GetMatchStatus(ctx context.Context, req *GetMatchStatusRequest, opts ...http.CallOption) (rsp *GetMatchStatusReply, err error)
	StartMatch(ctx context.Context, req *StartMatchRequest, opts ...http.CallOption) (rsp *StartMatchReply, err error)
}

type GameHTTPClientImpl struct {
	cc *http.Client
}

func NewGameHTTPClient(client *http.Client) GameHTTPClient {
	return &GameHTTPClientImpl{client}
}

func (c *GameHTTPClientImpl) CancelMatch(ctx context.Context, in *CancelMatchRequest, opts ...http.CallOption) (*CancelMatchReply, error) {
	var out CancelMatchReply
	pattern := "/v1/game/cancel"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationGameCancelMatch))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *GameHTTPClientImpl) GetMatchStatus(ctx context.Context, in *GetMatchStatusRequest, opts ...http.CallOption) (*GetMatchStatusReply, error) {
	var out GetMatchStatusReply
	pattern := "/v1/game/status/{matchid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGameGetMatchStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *GameHTTPClientImpl) StartMatch(ctx context.Context, in *StartMatchRequest, opts ...http.CallOption) (*StartMatchReply, error) {
	var out StartMatchReply
	pattern := "/v1/game/start"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationGameStartMatch))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}